Move directionApiResult to models package
