+++ app/src/main/java/io/github/vibrouter/MainService.java
@@ -35,6 +35,7 @@ import com.google.gson.GsonBuilder;
 import java.util.Collections;
 import java.util.List;
 
+import io.github.vibrouter.models.DirectionsApiResult;
 import io.github.vibrouter.utils.GpsUtil;
 
 public class MainService extends Service {
